// src/types/userType.tsx
export type userType = {
  area_code?: string;
  eamil?: string;
  name?: string;
  password?: string;
  check_password?: string;
  number?: string;
};

export type sellerType = {
  phone_number?: string;
  business_number?: string;
  username?: string;
  password?: string;
  check_password?: string;
  name?: string;
  email?: string;
  address?: string;
  call_number?: string;
  area_code?: string;
  bank_code?: string;
  bank_account?: string;
  bank_account_name?: string;
};